move Singleton to utils
